Update and rename ATMproblem.c to ATM-HS08.c
